requirements.txt
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
flasgger==0.9.7.1
Flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.2
mongoengine==0.27.0
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
WTForms==3.1.1
